change of the hosts reference
When all is used for hosts and no inventory file is present, ansible will create an implicit localhost, which is exactly what we want and need in this case. It means that the same playbook can be used an a development environment with Vagrant and VirtuaBox as in a GitHub Actions environment, without the need for something else.