Make it work for localhost
It seems when no inventory is used and the localhost is created implicitly, the current playbook does not work while hosts all, does not refer to localhost for the current setup. This uses the command line to define localhost for the GitHub case and the VirtuallBox machine when running the  playbook in the development environment.